requirements-dev.txt
-r requirements.txt
nose
rednose
coverage
django-nose
pytest
pytest-cov
pytest-django
flake8
tox
ansible
#boto
#fabric3
